requirements.txt
streamlit
langchain
chromadb
python-dotenv
langchain-huggingface
langchain-community

